Increase powers to 32bit data type
Some saves are reporting to have powers of up to 99999

+++ src/models.rs
@@ -530,7 +530,7 @@ pub struct Country {
     pub previous_monarchs: Vec<ObjId>,
     pub government: Option<CountryGovernment>,
     #[jomini(default)]
-    pub powers: [i16; 3],
+    pub powers: [i32; 3],
     #[jomini(default)]
     pub mercantilism: f32,
     #[jomini(default)]
